* 6.3:
  [Validator] Add missing validator translations in Polish language
  [HttpClient] Fix encoding some characters in query strings
  [HttpKernel] make RequestPayloadValueResolver:resolve() throw on variadic argument
  [SecurityBundle] Remove last usages of tag `security.remember_me_aware`
  [VarDumper] Dumping DateTime throws error if getTimezone is false
  Only update autoload_runtime.php when it changed
  [FrameworkBundle] Fix secrets:list not displaying local vars
  [Intl] Update the ICU data to 73.2
  [DoctrineBridge] add missing UPGRADE notes for #50689
  [HttpClient] Force int conversion for floated multiplier for GenericRetryStrategy
  [Security] Fix log message in OidcTokenHandler
  Don't mark RedispatchMessage as internal
  [FrameworkBundle] Ignore missing directories in about command
  Revert "[Messenger] Respect `isRetryable` decision of the retry strategy when deciding if failed message should be re-delivered"
  [Validator] Add the `message` option to the `PasswordStrength` constraint
  [Validator][Translator] Fix xlf files for en & fr translations. Bug introduced by #50590
  CS fix
  Add missing EN and FR translations for newest constraints
  [HttpClient] Remove final keyword on AsyncResponse
  [DependencyInjection] Fix support for `false` boolean env vars

+++ src/Symfony/Component/HttpClient/Retry/GenericRetryStrategy.php
@@ -102,7 +102,7 @@ class GenericRetryStrategy implements RetryStrategyInterface
         $delay = $this->delayMs * $this->multiplier ** $context->getInfo('retry_count');
 
         if ($this->jitter > 0) {
-            $randomness = $delay * $this->jitter;
+            $randomness = (int) ($delay * $this->jitter);
             $delay = $delay + random_int(-$randomness, +$randomness);
         }
